requirements.txt
flask>=2.3.0
pyjwt>=2.8.0
joblib>=1.3.0
numpy>=1.24.0
sqlalchemy>=2.0.0
scikit-learn>=1.3.0

